Separate out datemaytime type
It is nothing to do with Meta.hs
