+++ neuron/src/lib/Neuron/Zettelkasten/Zettel/Parser.hs
@@ -8,6 +8,7 @@ module Neuron.Zettelkasten.Zettel.Parser where
 import Control.Monad.Writer
 import Data.Some
 import qualified Data.Text as T
+import Data.Time.DateMayTime (mkDateMayTime)
 import Neuron.Reader.Type
 import Neuron.Zettelkasten.ID
 import Neuron.Zettelkasten.Query.Error
@@ -40,7 +41,7 @@ parseZettel format zreader fn zid s = do
           date = case zid of
             -- We ignore the "data" meta field on legacy Date IDs, which encode the
             -- creation date in the ID.
-            ZettelDateID v _ -> Just $ Meta.mkDateMayTime $ Left v
+            ZettelDateID v _ -> Just $ mkDateMayTime $ Left v
             ZettelCustomID _ -> Meta.date =<< meta
           unlisted = fromMaybe False $ Meta.unlisted =<< meta
           (queries, errors) = runWriter $ extractQueries doc
